Add files via upload
Implemented the concept of Exception Handling in Spring Security.
It contains a package named "exceptionhandling" which contains two clasess named as "CustomBasicAuthenticationEntryPoint" and "CustomAccessDeniedHandler".
CustomBasicAuthenticationEntryPoint is used to handle the the AuthenticationException which occurs due to authentication process failed or invalid i.e. (invalid username or password) by overiding the only method that is commence() method .
CustomAccessDeniedHandler is used to handle the AccessDeniedException which occurs when the user is authenticated but not authorized to access particular resource.